Dedup livemap demos
