+++ livemap-demo/src/jvmRawAwtMain/kotlin/jetbrains/livemap/demo/PerdRawAwtDemo.kt
@@ -1,14 +1,14 @@
 /*
- * Copyright (c) 2021. JetBrains s.r.o.
+ * Copyright (c) 2020. JetBrains s.r.o.
  * Use of this source code is governed by the MIT license that can be found in the LICENSE file.
  */
 
 package jetbrains.livemap.demo
 
-object BlankLiveMapDemoAwt {
-
+object PerfRawAwtDemo {
     @JvmStatic
     fun main(args: Array<String>) {
-        RawAwtDemo(::EmptyLiveMapDemoModel).start()
+        RawAwtDemo(::PerfDemoModel).start()
     }
 }
+
